UHF-6046: split panel and menu and widget toggle to separate files. Enhance langmenu with js button
