+++ src/scss/06_components/navigation/global/_menu-toggle.scss
@@ -206,10 +206,16 @@ $-menu-transition-duration: 0.3s;
 
 
   & ~ * .block__headerlanguagelinks,
-  
   & ~ * .headerlanguagelinks__wrapper
   {
     display:none;
+  }
+
+  
+  @include breakpoint($breakpoint-m) {
+    & ~ * .block__headerlanguagelinks {
+      display:flex;
+    }
   }
 
   & ~ * .menu-toggle__open {
@@ -248,7 +254,10 @@ $-menu-transition-duration: 0.3s;
     display: none;
   }
 }
-
+/**
+ * toggleWidgets.js could maybe inject this 
+ * instead of this hanging around in random file corner
+ */
 [data-cssmenu-hide] {
   display: none !important;
 }
