+++ templates/layout/page.html.twig
@@ -39,7 +39,7 @@
   {# This checkbox handles opening and closing mobile menu #}
   <input type="checkbox" id="cssmenu-toggle-checkbox" class="cssmenu-toggle-checkbox js-cssmenu-toggle-checkbox" role="button" aria-labelledby="cssmenu-toggle">
 {% endif %}
- <div id="morelangs"></div>
+ <div id="otherlangs"></div>
 <div{{ create_attribute( publishAttribute ).addClass(page_classes) }}>
   {% if page.header_top or page.header_bottom or page.header_branding %}
     <header role="banner" class="header">
